+++ Hawk/functions/User/Start-HawkUserInvestigation.ps1
@@ -142,9 +142,11 @@
 		if (Test-HawkGlobalObject) {
 			Initialize-HawkGlobalObject
 		}
+		$investigationStartTime = Get-Date
+
 
 		if ($PSCmdlet.ShouldProcess("Investigating Users")) {
-			Out-LogFile "Investigating Users" -Action
+			Out-LogFile "Starting User Investigation" -Action
 			Send-AIEvent -Event "CmdRun"
 	
 			# Pull the tenant configuration
@@ -175,10 +177,15 @@
 					Out-LogFile "Running Get-HawkUserAutoReply" -Action
 					Get-HawkUserAutoReply -User $User
 				}
+
+				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserEntraIDSignInLog for $User")) {
+					Out-LogFile "Running Get-HawkUserEntraIDSignInLog" -Action
+					Get-HawkUserEntraIDSignInLog -UserPrincipalName $User
+				}
 	
-				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserAuthHistory for $User")) {
-					Out-LogFile "Running Get-HawkUserAuthHistory" -Action
-					Get-HawkUserAuthHistory -User $User -ResolveIPLocations
+				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserUALSignInLog for $User")) {
+					Out-LogFile "Running Get-HawkUserUALSignInLog" -Action
+					Get-HawkUserUALSignInLog -User $User -ResolveIPLocations
 				}
 	
 				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserMailboxAuditing for $User")) {
@@ -200,7 +207,6 @@
 					Out-LogFile "Running Get-HawkUserMailItemsAccessed" -Action
 					Get-HawkUserMailItemsAccessed -UserPrincipalName $User
 				}
-
 				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserExchangeSearchQuery for $User")) {
 					Out-LogFile "Running Get-HawkUserExchangeSearchQuery" -Action
 					Get-HawkUserExchangeSearchQuery -UserPrincipalName $User
@@ -220,9 +226,18 @@
 					Out-LogFile "Running Get-HawkUserMobileDevice" -Action
 					Get-HawkUserMobileDevice -User $User
 				}
+
+				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserSharePointSearchQuery for $User")) {
+					Out-LogFile "Running Get-HawkUserSharePointSearchQuery" -Action
+					Get-HawkUserSharePointSearchQuery -UserPrincipalName $User
+				}
 			}
 		}
 
+	} end {
+        # Calculate end time and display summary
+        $investigationEndTime = Get-Date
+		Write-HawkInvestigationSummary -StartTime $investigationStartTime -EndTime $investigationEndTime -InvestigationType 'User' -UserPrincipalName $UserPrincipalName
 	}
 
 }